+++ GSClient/GSClient/SceneJH.h
@@ -93,7 +93,7 @@ public:
 
 	void ReleaseObjects(); 
 public:
-	virtual void Update(double elapsedTime) override;
+	virtual void Update(float elapsedTime) override;
 	void AnimateObjects(float fTimeElapsed);
 
 	virtual void Draw(ID3D12GraphicsCommandList* pd3dCommandList) override;
